Split Jira and Bugzilla services into separate modules
